Reorder functions if -f flag provided in cmd
Note: This commit contains a bug. It doesn't reorder
the functions as expected. It should only keep the
functions that are in the order file. But it reorders
the functions and merge them with functions that are not
in the order file. For example:

Functions in ELF:
-----------------

hello
gello
tello
mello

Functions in order file:
------------------------

tello
hello

Expected ELF:
-------------

tello
hello

Actual Output:
--------------

tello
gello
hello
mello